==> analysis/03_Deconvolution.Rmd <==
---
title: "Deconvolution file preparation"
output: 
  github_document:
     toc: true
     toc_depth: 3
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r include=FALSE}
library(DESeq2)
library(stringr)
library(dplyr)
library(purrr)
library(here)
library(biomaRt)
library(tidyr)
library(ggplot2)
library(cowplot)
```

# CIBERSORT

CIBERSORT requires TPM as input, therefore we need to convert our batch corrected counts to TPM. To get the gene lengths for this, we will use the effective gene lengths from the Salmon output. Therefore, we must use the ENGS gene id version.

## Generate inputs to CIBERSORT

The function `tpm.convert` is used to convert the counts to TPM.

```{r}
tpm.convert <- function(counts.mat, gene.length){
x <- counts.mat / gene.length
#Then with this matrix x, you do the following:
tpm.mat <- t( t(x) * 1e6 / colSums(x) )
}

read_in_quant_genes <- function(file.name){
  x <- read.table(file = file.name, header = T)[,c(1,3)]
  #colnames(x)[2] <- as.character(sample.name)
  return(x)
}
```

```{r}
#dds.ensg.remove.outliers.batch.corrected <- readRDS("../intermediate_files/dds_batch_corrected_group_tumor_ensembl_gene_id_version_2022-10-03.Rds")
dds.ensg.remove.outliers.batch.corrected <- readRDS("../intermediate_files/dds_batch_corrected_group_tumor_ensembl_gene_id_version_2022-10-07.Rds")
```

```{r}
metadata <- read.table(here("intermediate_files/metadata/metadata_all_samples.txt"), row.names = 1, sep = "\t")
files.gene <- file.path(metadata$directory, rownames(metadata), "quant.genes.sf")
files.gene.samples <- str_split_fixed(files.gene, pattern = "/", n = 11)[,10]
files.gene.outliers.removed <- files.gene[files.gene.samples %in% colnames(dds.ensg.remove.outliers.batch.corrected)]
sample.names.outliers.removed <- files.gene.samples[files.gene.samples %in% colnames(dds.ensg.remove.outliers.batch.corrected)]
```

```{r}
genes.effective.lengths <- lapply(files.gene.outliers.removed, read_in_quant_genes)
length(genes.effective.lengths)

genes.effective.lengths.list <- list()
for (i in 1:length(genes.effective.lengths)){
  df <- genes.effective.lengths[[i]]
  colnames(df)[2] <- sample.names.outliers.removed[i]
  genes.effective.lengths.list[[i]] <- df
}

```

```{r}
genes.effective.lengths.df <- genes.effective.lengths.list %>% purrr::reduce(inner_join, by = "Name")
mart <- useMart(biomart = "ensembl", dataset = "hsapiens_gene_ensembl", #host="https://www.ensembl.org")
                host="useast.ensembl.org")
tx2gene.ensg <- getBM(attributes = c("hgnc_symbol", "ensembl_gene_id_version"), mart = mart, useCache = FALSE)
colnames(tx2gene.ensg)[2] <- "Name"
genes.effective.lengths.df.hgnc <- left_join(genes.effective.lengths.df, tx2gene.ensg,by = "Name")
```

```{r}
df.dds.ensg.remove.outliers.batch.corrected <- assay(dds.ensg.remove.outliers.batch.corrected)[2:nrow(assay(dds.ensg.remove.outliers.batch.corrected)),]
dim(df.dds.ensg.remove.outliers.batch.corrected)
```

```{r}
rownames(genes.effective.lengths.df) <- NULL
genes.effective.lengths.df.filtered <- genes.effective.lengths.df[which(genes.effective.lengths.df$Name %in% rownames(df.dds.ensg.remove.outliers.batch.corrected)),]
dim(genes.effective.lengths.df.filtered)
df.dds.ensg.remove.outliers.batch.corrected.matched <- df.dds.ensg.remove.outliers.batch.corrected[genes.effective.lengths.df.filtered$Name,]
```

```{r}
rownames(genes.effective.lengths.df.filtered) <- genes.effective.lengths.df.filtered$Name
genes.effective.lengths.df.filtered$Name <- NULL
df.dds.ensg.remove.outliers.batch.corrected.matched <- data.frame(df.dds.ensg.remove.outliers.batch.corrected.matched)
genes.effective.lengths.df.filtered[1:5,1:5]
df.dds.ensg.remove.outliers.batch.corrected.matched[1:5,1:5]
```

```{r}
tpm.out <- tpm.convert(counts.mat =df.dds.ensg.remove.outliers.batch.corrected.matched, gene.length = genes.effective.lengths.df.filtered )
tpm.out[1:5,1:5]
```

```{r TPM data for mixture matrix}
tpm.out.mixture <- data.frame(tpm.out[,which(colData(dds.ensg.remove.outliers.batch.corrected)$Study == "InHouse")])
tpm.out.mixture.genes <- tibble::rownames_to_column(tpm.out.mixture, "genes")
tpm.out.mixture.genes[1:5,1:5]
tpm.out.mixture.genes.outfile <- paste("intermediate_files/cibersort/caf_tpm_mixture_engs_version_batch_corrected_", Sys.Date(), ".txt", sep = "")
#write.table(x = tpm.out.mixture.genes, file = here(tpm.out.mixture.genes.outfile), sep = "\t", quote = F, row.names = F)
```

```{r TPM data for signature matrix}
# signature is our non-inhouse
tpm.out.signature <- tpm.out[,which(dds.ensg.remove.outliers.batch.corrected$Study != "InHouse")]
colnames(tpm.out.signature) <- dds.ensg.remove.outliers.batch.corrected$Subpopulation[which(dds.ensg.remove.outliers.batch.corrected$Study != "InHouse")]
tpm.out.signature[1:5,1:5]
```

```{r}
s1_cols <- which(colnames(tpm.out.signature) == "S1")
s3_cols <- which(colnames(tpm.out.signature) == "S3")
s4_cols <- which(colnames(tpm.out.signature) == "S4")

s1_cols_length <- length(s1_cols)
s3_cols_length <- length(s3_cols)
s4_cols_length <- length(s4_cols)
```

```{r Put signature data in the right order ie all the S1s together etc}
tpm.data.no.inhouse.order <- cbind.data.frame(rownames(tpm.out.signature),
                                              tpm.out.signature[,s1_cols],
                                              tpm.out.signature[,s3_cols],
                                              tpm.out.signature[,s4_cols]
                                              )
colnames(tpm.data.no.inhouse.order) <- NULL
rownames(tpm.data.no.inhouse.order) <- NULL
cols_no_inhouse <- c("genes", rep("S1", s1_cols_length), rep("S3", s3_cols_length), rep("S4", s4_cols_length))
tpm.data.no.inhouse.order <- rbind.data.frame(cols_no_inhouse, tpm.data.no.inhouse.order)
tpm.data.no.inhouse.order[1:5,1:5]
tpm.data.no.inhouse.order.outfile <- paste("intermediate_files/cibersort/caf_tpm_for_signature_engs_version_batch_corrected_", Sys.Date(), ".txt", sep = "")
#write.table(tpm.data.no.inhouse.order, file = here(tpm.data.no.inhouse.order.outfile), sep = "\t", quote = F, row.names = F, col.names = F)
```



```{r create phenoclasses file}
S1 <- c(rep(1,s1_cols_length), rep(2, (s3_cols_length+s4_cols_length)))
S3 <- c(rep(2,s1_cols_length), rep(1, s3_cols_length), rep(2, s4_cols_length))
S4 <- c(rep(2, (s1_cols_length+s3_cols_length)), rep(1, s4_cols_length))


if (length(S1) != length(colnames(tpm.out.signature))){
  stop("length s1 row incorrect")
} else if (length(S3) != length(colnames(tpm.out.signature))) {
  stop("length s3 row incorrect")
} else if (length(S4) != length(colnames(tpm.out.signature))) {
  stop("length s4 row incorrect")
}

out <- rbind.data.frame(S1, S3, S4)
colnames(out) <- NULL
row.names(out) <- c("S1", "S3", "S4")
outfile.name <-  paste("intermediate_files/cibersort/phenoclasses_caf_", Sys.Date(), ".txt", sep = "")
#write.table(out, file = here(outfile.name), sep = "\t", quote = F, col.names = F)
```
